+++ app/charts/shared/possible-filters.ts
